ruleUtils domUtils getFieldConfig - config is 1st param
