+++ modules/components/widgets/antd/core/FieldTreeSelect.jsx
@@ -1,6 +1,6 @@
 import React, { PureComponent } from "react";
 import { Tooltip, TreeSelect } from "antd";
-import {BUILT_IN_PLACEMENTS, SELECT_WIDTH_OFFSET_RIGHT, calcTextWidth} from "../../../../utils/stuff";
+import {BUILT_IN_PLACEMENTS, SELECT_WIDTH_OFFSET_RIGHT, calcTextWidth} from "../../../../utils/domUtils";
 import {useOnPropsChanged} from "../../../../utils/reactUtils";
 import PropTypes from "prop-types";
 import keys from "lodash/keys";
