+++ modules/export/jsonLogic.js
@@ -1,8 +1,8 @@
-"use strict";
 import {defaultValue} from "../utils/stuff";
 import {
-  getFieldConfig, getWidgetForFieldOp, getOperatorConfig, getFieldWidgetConfig, getFuncConfig
+  getFieldConfig, getOperatorConfig, getFieldWidgetConfig, getFuncConfig
 } from "../utils/configUtils";
+import {getWidgetForFieldOp} from "../utils/ruleUtils";
 import {defaultConjunction} from "../utils/defaultUtils";
 import {completeValue} from "../utils/funcUtils";
 import {Map} from "immutable";
@@ -25,12 +25,12 @@ export const jsonLogicFormat = (item, config) => {
   const {fieldSeparator} = config.settings;
   let data = {};
   for (let ff of usedFields) {
-    const def = getFieldConfig(ff, config) || {};
+    const def = getFieldConfig(config, ff) || {};
     const parts = ff.split(fieldSeparator);
     let tmp = data;
     for (let i = 0 ; i < parts.length ; i++) {
       const p = parts[i];
-      const pdef = getFieldConfig(parts.slice(0, i+1), config) || {};
+      const pdef = getFieldConfig(config, parts.slice(0, i+1)) || {};
       if (i != parts.length - 1) {
         if (pdef.type == "!group" && pdef.mode != "struct") {
           if (!tmp[p])
@@ -59,7 +59,7 @@ export const jsonLogicFormat = (item, config) => {
 const formatFunc = (meta, config, currentValue, parentField = null) => {
   const funcKey = currentValue.get("func");
   const args = currentValue.get("args");
-  const funcConfig = getFuncConfig(funcKey, config);
+  const funcConfig = getFuncConfig(config, funcKey);
   if (!funcConfig.jsonLogic) {
     meta.errors.push(`Func ${funcKey} is not supported`);
     return undefined;
@@ -67,7 +67,7 @@ const formatFunc = (meta, config, currentValue, parentField = null) => {
   const formattedArgs = {};
   for (const argKey in funcConfig.args) {
     const argConfig = funcConfig.args[argKey];
-    const fieldDef = getFieldConfig(argConfig, config);
+    const fieldDef = getFieldConfig(config, argConfig);
     const argVal = args ? args.get(argKey) : undefined;
     const argValue = argVal ? argVal.get("value") : undefined;
     const argValueSrc = argVal ? argVal.get("valueSrc") : undefined;
@@ -113,13 +113,13 @@ const formatField = (meta, config, field, parentField = null) => {
   if (field) {
     if (Array.isArray(field))
       field = field.join(fieldSeparator);
-    const fieldDef = getFieldConfig(field, config) || {};
+    const fieldDef = getFieldConfig(config, field) || {};
     let fieldName = field;
     if (fieldDef.fieldName) {
       fieldName = fieldDef.fieldName;
     }
     if (parentField) {
-      const parentFieldDef = getFieldConfig(parentField, config) || {};
+      const parentFieldDef = getFieldConfig(config, parentField) || {};
       let parentFieldName = parentField;
       if (parentFieldDef.fieldName) {
         parentFieldName = parentFieldDef.fieldName;
@@ -169,7 +169,7 @@ const formatValue = (meta, config, currentValue, valueSrc, valueType, fieldWidge
 
 const formatItemValue = (config, properties, meta, operator, parentField) => {
   const field = properties.get("field");
-  const fieldDefinition = getFieldConfig(field, config) || {};
+  const fieldDefinition = getFieldConfig(config, field) || {};
   const operatorDefinition = getOperatorConfig(config, operator, field) || {};
   const cardinality = defaultValue(operatorDefinition.cardinality, 1);
 
@@ -341,7 +341,7 @@ const formatRule = (item, config, meta, parentField = null) => {
   if (field == null || operator == null)
     return undefined;
 
-  const fieldDefinition = getFieldConfig(field, config) || {};
+  const fieldDefinition = getFieldConfig(config, field) || {};
   let operatorDefinition = getOperatorConfig(config, operator, field) || {};
   let reversedOp = operatorDefinition.reversedOp;
   let revOperatorDefinition = getOperatorConfig(config, reversedOp, field) || {};
