+++ modules/export/queryString.js
@@ -1,6 +1,7 @@
 import {
-  getFieldConfig, getWidgetForFieldOp, getOperatorConfig, getFieldWidgetConfig, getFieldPath, getFieldPathLabels, getFuncConfig
+  getFieldConfig, getOperatorConfig, getFieldWidgetConfig, getFuncConfig
 } from "../utils/configUtils";
+import {getFieldPath, getFieldPathLabels, getWidgetForFieldOp} from "../utils/ruleUtils";
 import omit from "lodash/omit";
 import pick from "lodash/pick";
 import {defaultValue} from "../utils/stuff";
@@ -10,7 +11,7 @@ import {completeValue} from "../utils/funcUtils";
 import {Map} from "immutable";
 
 const formatFieldName = (field, config) => {
-  const fieldDefinition = getFieldConfig(field, config) || {};
+  const fieldDefinition = getFieldConfig(config, field) || {};
   const {fieldSeparator} = config.settings;
   const fieldParts = Array.isArray(field) ? field : field.split(fieldSeparator);
   let fieldName = Array.isArray(field) ? field.join(fieldSeparator) : field;
@@ -33,7 +34,7 @@ const formatValue = (config, currentValue, valueSrc, valueType, fieldWidgetDefin
     const rightField = currentValue;
     let formattedField = null;
     if (rightField) {
-      const rightFieldDefinition = getFieldConfig(rightField, config) || {};
+      const rightFieldDefinition = getFieldConfig(config, rightField) || {};
       const fieldParts = Array.isArray(rightField) ? rightField : rightField.split(fieldSeparator);
       const _fieldKeys = getFieldPath(rightField, config);
       const fieldPartsLabels = getFieldPathLabels(rightField, config);
@@ -47,13 +48,13 @@ const formatValue = (config, currentValue, valueSrc, valueType, fieldWidgetDefin
   } else if (valueSrc == "func") {
     const funcKey = currentValue.get("func");
     const args = currentValue.get("args");
-    const funcConfig = getFuncConfig(funcKey, config);
+    const funcConfig = getFuncConfig(config, funcKey);
     const funcName = isForDisplay && funcConfig.label || funcKey;
     const formattedArgs = {};
     const formattedArgsWithNames = {};
     for (const argKey in funcConfig.args) {
       const argConfig = funcConfig.args[argKey];
-      const fieldDef = getFieldConfig(argConfig, config);
+      const fieldDef = getFieldConfig(config, argConfig);
       const argVal = args ? args.get(argKey) : undefined;
       const argValue = argVal ? argVal.get("value") : undefined;
       const argValueSrc = argVal ? argVal.get("valueSrc") : undefined;
@@ -89,7 +90,7 @@ const formatValue = (config, currentValue, valueSrc, valueType, fieldWidgetDefin
         args.push(operatorDefinition);
       }
       if (valueSrc == "field") {
-        const valFieldDefinition = getFieldConfig(currentValue, config) || {}; 
+        const valFieldDefinition = getFieldConfig(config, currentValue) || {}; 
         args.push(valFieldDefinition);
       }
       ret = fn(...args);
@@ -127,7 +128,7 @@ export const queryString = (item, config, isForDisplay = false) => {
     if (field == null || operator == null)
       return undefined;
 
-    const fieldDefinition = getFieldConfig(field, config) || {};
+    const fieldDefinition = getFieldConfig(config, field) || {};
     const operatorDefinition = getOperatorConfig(config, operator, field) || {};
     const reversedOp = operatorDefinition.reversedOp;
     const revOperatorDefinition = getOperatorConfig(config, reversedOp, field) || {};
