+++ modules/import/index.js
@@ -1,4 +1,2 @@
-"use strict";
-
 export * from "./tree";
 export * from "./jsonLogic";
