+++ modules/import/jsonLogic.js
@@ -1,6 +1,7 @@
 import uuid from "../utils/uuid";
 import {defaultValue} from "../utils/stuff";
-import {getFieldConfig, extendConfig, getWidgetForFieldOp} from "../utils/configUtils";
+import {getFieldConfig, extendConfig} from "../utils/configUtils";
+import {getWidgetForFieldOp} from "../utils/ruleUtils";
 import {loadTree} from "./tree";
 import {defaultConjunction, defaultGroupConjunction} from "../utils/defaultUtils";
 import moment from "moment";
@@ -177,7 +178,7 @@ const convertField = (op, vals, conv, config, not, meta, parentField = null) => 
     let field = vals[0];
     if (parentField)
       field = [parentField, field].join(fieldSeparator);
-    const fieldConfig = getFieldConfig(field, config); //todo: use fieldName config for reverse find (introduce cache?)
+    const fieldConfig = getFieldConfig(config, field); //todo: use fieldName config for reverse find (introduce cache?)
     if (!fieldConfig) {
       meta.errors.push(`No config for field ${field}`);
       return undefined;
@@ -265,7 +266,7 @@ const convertConj = (op, vals, conv, config, not, meta, parentField = null) => {
     const complexFieldsParents = complexFields
       .map(parts => parts.slice(0, parts.length - 1).join(fieldSeparator));
     const complexFieldsConfigs = arrayToObject(
-      arrayUniq(complexFieldsParents).map(f => [f, getFieldConfig(f, config)])
+      arrayUniq(complexFieldsParents).map(f => [f, getFieldConfig(config, f)])
     );
     const complexFieldsInRuleGroup = complexFieldsParents
       .filter((f) => complexFieldsConfigs[f].type == "!group");
@@ -300,7 +301,7 @@ const convertConj = (op, vals, conv, config, not, meta, parentField = null) => {
           groupField.split(fieldSeparator).map((f, i, a) => {
             const p = a.slice(0, i);
             const ff = [...p, f].join(fieldSeparator);
-            const ffConfig = getFieldConfig(ff, config) || {};
+            const ffConfig = getFieldConfig(config, ff) || {};
             if (!needSplit && i == 0) {
               type = "rule_group";
               properties.field = ff;
@@ -466,7 +467,7 @@ const _parseRule = (op, arity, vals, parentField, conv, config, errors, isRevArg
     if (parentField)
       field = [parentField, field].join(fieldSeparator);
 
-    const fieldConfig = getFieldConfig(field, config);
+    const fieldConfig = getFieldConfig(config, field);
     if (!fieldConfig) {
       errors.push(`No config for field ${field}`);
       return;
