+++ modules/utils/configUtils.js
@@ -1,10 +1,9 @@
-"use strict";
-import last from "lodash/last";
 import merge from "lodash/merge";
 import mergeWith from "lodash/mergeWith";
 import {settings as defaultSettings} from "../config/default";
 import moment from "moment";
 import {normalizeListValues} from "./stuff";
+import {getWidgetForFieldOp} from "./ruleUtils";
 
 
 export const extendConfig = (config) => {
@@ -23,7 +22,7 @@ export const extendConfig = (config) => {
   _extendFieldsConfig(config.fields, config);
 
   _extendFuncArgsConfig(config.funcs, config);
-    
+
   moment.locale(config.settings.locale.moment);
 
   Object.defineProperty(config, "__extended", {
@@ -183,7 +182,7 @@ function _extendFieldConfig(fieldConfig, config, isFuncArg = false) {
   }
 }
 
-export const getFieldRawConfig = (field, config, fieldsKey = "fields", subfieldsKey = "subfields") => {
+export const getFieldRawConfig = (config, field, fieldsKey = "fields", subfieldsKey = "subfields") => {
   if (!field)
     return null;
   const fieldSeparator = config.settings.fieldSeparator;
@@ -214,17 +213,17 @@ export const getFieldRawConfig = (field, config, fieldsKey = "fields", subfields
 };
 
 
-export const getFuncConfig = (func, config) => {
+export const getFuncConfig = (config, func) => {
   if (!func)
     return null;
-  const funcConfig = getFieldRawConfig(func, config, "funcs", "subfields");
+  const funcConfig = getFieldRawConfig(config, func, "funcs", "subfields");
   if (!funcConfig)
     return null; //throw new Error("Can't find func " + func + ", please check your config");
   return funcConfig;
 };
 
-export const getFuncArgConfig = (funcKey, argKey, config) => {
-  const funcConfig = getFuncConfig(funcKey, config);
+export const getFuncArgConfig = (config, funcKey, argKey) => {
+  const funcConfig = getFuncConfig(config, funcKey);
   if (!funcConfig)
     return null; //throw new Error(`Can't find func ${funcKey}, please check your config`);
   const argConfig = funcConfig.args && funcConfig.args[argKey] || null;
@@ -242,14 +241,14 @@ export const getFuncArgConfig = (funcKey, argKey, config) => {
   return ret;
 };
 
-export const getFieldConfig = (field, config) => {
+export const getFieldConfig = (config, field) => {
   if (!field)
     return null;
   if (typeof field == "object" && !field.func && !!field.type)
     return field;
   if (typeof field == "object" && field.func && field.arg)
-    return getFuncArgConfig(field.func, field.arg, config);
-  const fieldConfig = getFieldRawConfig(field, config);
+    return getFuncArgConfig(config, field.func, field.arg);
+  const fieldConfig = getFieldRawConfig(config, field);
   if (!fieldConfig)
     return null; //throw new Error("Can't find field " + field + ", please check your config");
 
@@ -264,74 +263,12 @@ export const getFieldConfig = (field, config) => {
   return ret;
 };
 
-export const getFirstField = (config, parentRuleGroupPath = null) => {
-  const fieldSeparator = config.settings.fieldSeparator;
-  const parentPathArr = typeof parentRuleGroupPath == "string" ? parentRuleGroupPath.split(fieldSeparator) : parentRuleGroupPath;
-  const parentField = parentRuleGroupPath ? getFieldRawConfig(parentRuleGroupPath, config) : config;
-
-  let firstField = parentField, key = null, keysPath = [];
-  do {
-    const subfields = firstField === config ? config.fields : firstField.subfields;
-    if (!subfields || !Object.keys(subfields).length) {
-      firstField = key = null;
-      break;
-    }
-    key = Object.keys(subfields)[0];
-    keysPath.push(key);
-    firstField = subfields[key];
-  } while (firstField.type == "!struct" || firstField.type == "!group");
-  return (parentPathArr || []).concat(keysPath).join(fieldSeparator);
-};
-
-export const getOperatorsForField = (config, field) => {
-  const fieldConfig = getFieldConfig(field, config);
-  const fieldOps = fieldConfig ? fieldConfig.operators : [];
-  return fieldOps;
-};
-
-export const getFirstOperator = (config, field) => {
-  const fieldOps = getOperatorsForField(config, field);
-  return fieldOps ? fieldOps[0] : null;
-};
-
-export const getFieldPath = (field, config, onlyKeys = false) => {
-  if (!field)
-    return null;
-  const fieldSeparator = config.settings.fieldSeparator;
-  const parts = Array.isArray(field) ? field : field.split(fieldSeparator);
-  if (onlyKeys)
-    return parts;
-  else
-    return parts
-      .map((_curr, ind, arr) => arr.slice(0, ind+1))
-      .map((parts) => parts.join(fieldSeparator));
-};
-
-export const getFuncPathLabels = (field, config) => {
-  return getFieldPathLabels(field, config, "funcs", "subfields");
-};
-
-export const getFieldPathLabels = (field, config, fieldsKey = "fields", subfieldsKey = "subfields") => {
-  if (!field)
-    return null;
-  const fieldSeparator = config.settings.fieldSeparator;
-  const parts = Array.isArray(field) ? field : field.split(fieldSeparator);
-  return parts
-    .map((_curr, ind, arr) => arr.slice(0, ind+1))
-    .map((parts) => parts.join(fieldSeparator))
-    .map(part => {
-      const cnf = getFieldRawConfig(part, config, fieldsKey, subfieldsKey);
-      return cnf && cnf.label || cnf && last(part.split(fieldSeparator));
-    })
-    .filter(label => label != null);
-};
-
 export const getOperatorConfig = (config, operator, field = null) => {
   if (!operator)
     return null;
   const opConfig = config.operators[operator];
   if (field) {
-    const fieldConfig = getFieldConfig(field, config);
+    const fieldConfig = getFieldConfig(config, field);
     const widget = getWidgetForFieldOp(config, field, operator);
     const widgetConfig = config.widgets[widget] || {};
     const fieldWidgetConfig = (fieldConfig && fieldConfig.widgets ? fieldConfig.widgets[widget] : {}) || {};
@@ -347,7 +284,7 @@ export const getOperatorConfig = (config, operator, field = null) => {
 export const getFieldWidgetConfig = (config, field, operator, widget = null, valueSrc = null) => {
   if (!field || !(operator || widget))
     return null;
-  const fieldConfig = getFieldConfig(field, config);
+  const fieldConfig = getFieldConfig(config, field);
   if (!widget)
     widget = getWidgetForFieldOp(config, field, operator, valueSrc);
   const widgetConfig = config.widgets[widget] || {};
@@ -356,118 +293,4 @@ export const getFieldWidgetConfig = (config, field, operator, widget = null, val
   const valueFieldSettings = (valueSrc == "value" || !valueSrc) && fieldConfig && fieldConfig.fieldSettings || {}; // useful to take 'validateValue'
   const mergedConfig = merge({}, widgetConfig, fieldWidgetProps, valueFieldSettings);
   return mergedConfig;
-};
-
-export const getValueLabel = (config, field, operator, delta, valueSrc = null, isSpecialRange = false) => {
-  const isFuncArg = typeof field == "object" && field.arg;
-  const {showLabels} = config.settings;
-  const fieldConfig = getFieldConfig(field, config);
-  const fieldWidgetConfig = getFieldWidgetConfig(config, field, operator, null, valueSrc) || {};
-  const mergedOpConfig = getOperatorConfig(config, operator, field) || {};
-    
-  const cardinality = isSpecialRange ? 1 : mergedOpConfig.cardinality;
-  let ret = null;
-  if (cardinality > 1) {
-    const valueLabels = fieldWidgetConfig.valueLabels || mergedOpConfig.valueLabels;
-    if (valueLabels)
-      ret = valueLabels[delta];
-    if (ret && typeof ret != "object") {
-      ret = {label: ret, placeholder: ret};
-    }
-    if (!ret) {
-      ret = {
-        label: config.settings.valueLabel + " " + (delta+1),
-        placeholder: config.settings.valuePlaceholder + " " + (delta+1),
-      };
-    }
-  } else {
-    let label = fieldWidgetConfig.valueLabel;
-    let placeholder = fieldWidgetConfig.valuePlaceholder;
-    if (isFuncArg) {
-      if (!label)
-        label = fieldConfig.label || field.arg;
-      if (!placeholder && !showLabels)
-        placeholder = fieldConfig.label || field.arg;
-    }
-
-    ret = {
-      label: label || config.settings.valueLabel, 
-      placeholder: placeholder || config.settings.valuePlaceholder,
-    };
-  }
-  return ret;
-};
-
-function _getWidgetsAndSrcsForFieldOp (config, field, operator = null, valueSrc = null) {
-  let widgets = [];
-  let valueSrcs = [];
-  if (!field)
-    return {widgets, valueSrcs};
-  const isFuncArg = typeof field == "object" && !!field.type;
-  const fieldConfig = getFieldConfig(field, config);
-  const opConfig = operator ? config.operators[operator] : null;
-  if (fieldConfig && fieldConfig.widgets) {
-    for (let widget in fieldConfig.widgets) {
-      const widgetConfig = fieldConfig.widgets[widget];
-      const widgetValueSrc = config.widgets[widget].valueSrc || "value";
-      let canAdd = true;
-      if (!widgetConfig.operators)
-        canAdd = canAdd && (valueSrc != "value" || isFuncArg); //if can't check operators, don't add
-      if (widgetConfig.operators && operator)
-        canAdd = canAdd && widgetConfig.operators.indexOf(operator) != -1;
-      if (valueSrc && valueSrc != widgetValueSrc)
-        canAdd = false;
-      if (opConfig && opConfig.cardinality == 0 && (widgetValueSrc != "value"))
-        canAdd = false;
-      if (canAdd) {
-        widgets.push(widget);
-        let canAddValueSrc = fieldConfig.valueSources && fieldConfig.valueSources.indexOf(widgetValueSrc) != -1;
-        if (opConfig && opConfig.valueSources && opConfig.valueSources.indexOf(widgetValueSrc) == -1)
-          canAddValueSrc = false;
-        if (canAddValueSrc && !valueSrcs.find(v => v == widgetValueSrc))
-          valueSrcs.push(widgetValueSrc);
-      }
-    }
-  }
-  widgets.sort((w1, w2) => {
-    let w1Main = fieldConfig.preferWidgets ? fieldConfig.preferWidgets.indexOf(w1) != -1 : w1 == fieldConfig.mainWidget;
-    let _w2Main = fieldConfig.preferWidgets ? fieldConfig.preferWidgets.indexOf(w2) != -1 : w2 == fieldConfig.mainWidget;
-    if (w1 != w2) {
-      return w1Main ? -1 : +1;
-    }
-    return 0;
-  });
-    
-  return {widgets, valueSrcs};
-}
-
-export const getWidgetsForFieldOp = (config, field, operator, valueSrc = null) => {
-  const {widgets} = _getWidgetsAndSrcsForFieldOp(config, field, operator, valueSrc);
-  return widgets;
-};
-
-export const getValueSourcesForFieldOp = (config, field, operator, fieldDefinition = null, leftFieldForFunc = null) => {
-  const {valueSrcs} = _getWidgetsAndSrcsForFieldOp(config, field, operator, null);
-  const filteredValueSrcs = valueSrcs.filter(vs => {
-    if (vs == "field" && fieldDefinition) {
-      return config._fieldsCntByType[fieldDefinition.type] > 1;
-    }
-    if (vs == "func" && fieldDefinition) {
-      if (!config._funcsCntByType[fieldDefinition.type])
-        return false;
-      if (fieldDefinition.funcs)
-        return fieldDefinition.funcs.length > 0;
-      return true;
-    }
-    return true;
-  });
-  return filteredValueSrcs;
-};
-
-export const getWidgetForFieldOp = (config, field, operator, valueSrc = null) => {
-  const {widgets} = _getWidgetsAndSrcsForFieldOp(config, field, operator, valueSrc);
-  let widget = null;
-  if (widgets.length)
-    widget = widgets[0];
-  return widget;
 };
