+++ modules/utils/index.js
@@ -1,3 +1,2 @@
-"use strict";
 export {validateTree} from "./validation";
 export {default as uuid} from "./uuid";
