+++ modules/utils/stuff.js
@@ -1,42 +1,11 @@
-import mapValues from "lodash/mapValues";
 import Immutable, { Map } from "immutable";
-import React from "react";
-
-
-export const SELECT_WIDTH_OFFSET_RIGHT = 48;
-const DEFAULT_FONT_SIZE = "14px";
-const DEFAULT_FONT_FAMILY = "'Helvetica Neue', Helvetica, Arial, sans-serif";
 
 // RegExp.quote = function (str) {
 //     return str.replace(/([.?*+^$[\]\\(){}|-])/g, "\\$1");
 // };
 
-
 export const defaultValue = (value, _default) => {
   return (typeof value === "undefined") ? _default : value;
-};
-
-
-export const bindActionCreators = (actionCreators, config, dispatch) =>
-  mapValues(actionCreators, (actionCreator) =>
-    (...args) => dispatch(actionCreator(config, ...args))
-  );
-
-
-export const calcTextWidth = function(str, fontFamily = DEFAULT_FONT_FAMILY, fontSize = DEFAULT_FONT_SIZE) {
-  var div = document.createElement("div");
-  div.innerHTML = str;
-  var css = {
-    "position": "absolute", "float": "left", "white-space": "nowrap", "visibility": "hidden", 
-    "font-size": fontSize, "font-family": fontFamily
-  };
-  for (let k in css) {
-    div.style[k] = css[k];
-  }
-  div = document.body.appendChild(div);
-  var w = div.offsetWidth;
-  document.body.removeChild(div);
-  return w;
 };
 
 export const truncateString = (str, n, useWordBoundary) => {
@@ -45,41 +14,6 @@ export const truncateString = (str, n, useWordBoundary) => {
   return (useWordBoundary 
     ? subString.substr(0, subString.lastIndexOf(" ")) 
     : subString) + "...";
-};
-
-export const BUILT_IN_PLACEMENTS = {
-  bottomLeft: {
-    points: ["tl", "bl"],
-    offset: [0, 4],
-    overflow: {
-      adjustX: 0,
-      adjustY: 1,
-    },
-  },
-  bottomRight: {
-    points: ["tr", "br"],
-    offset: [0, 4],
-    overflow: {
-      adjustX: 1,
-      adjustY: 1,
-    },
-  },
-  topLeft: {
-    points: ["bl", "tl"],
-    offset: [0, -4],
-    overflow: {
-      adjustX: 0,
-      adjustY: 1,
-    },
-  },
-  topRight: {
-    points: ["br", "tr"],
-    offset: [0, -4],
-    overflow: {
-      adjustX: 1,
-      adjustY: 1,
-    },
-  },
 };
 
 export const immutableEqual = function(v1, v2) {
